requirements.txt
pytest~=7.1.2
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
requests~=2.26.0
PyYAML~=6.0
selenium~=4.1.3